Customer Information System
This program is written in C language, and it uses functions, matrices, and loops. It is designed for hairdressers to keep track of their clients' hair coloring history at a basic level. This console application can register customers, display the past dye numbers of registered customers, delete customer records, and add new dye numbers. It provides the authorized user with the ability to access client information and offer the most accurate service.